bug(asset detail): fix failing tests  - fix failing tests
